refactor(svelte): extract surface input handlers from surface-state
Move pointer/keyboard/capture side-effect switches and shared brush helpers
into surface-handlers.ts with live bindings for tool/brush state. Keep
reducer construction, deriveds, and phased effects in surface-state.svelte.ts
so construction guards and effect order stay load-bearing (compatible with
the #491 inspect pointer-frame protocol).

Stabilize themes overflow journeys by waiting for the specimen list before
asserting scrollWidth. Closes #467.
